muita coisa

+++ src/php/deletaUsuario.php
@@ -22,9 +22,19 @@ if($_SERVER["REQUEST_METHOD"] == "POST"){
 
     }elseif($nivelAcess == 4){
 
-
-        $sql1 ="DELETE FROM img_veiculo WHERE idVeiculos = '$idUser'";
+        $sql2 = "SELECT * FROM img_veiculo WHERE idVeiculos = '$idUser'";
+        $resultado2 = $conn->query($sql2);
+        if($resultado2->num_rows > 0){
+            while($row = $resultado2->fetch_assoc()){
+                $filename = "../".$row["caminho_imagem"];
+                if(file_exists($filename)){
+                    unlink($filename);
+                }
+            }
+            $sql1 ="DELETE FROM img_veiculo WHERE idVeiculos = '$idUser'";
         $conn->query($sql1);
+        }
+        
         $sql = "DELETE FROM veiculos WHERE id_veiculos = $idUser";
 
         echo "4";
